+++ src/js/control-bar/progress-control/seek-bar.js
@@ -2,7 +2,6 @@ import Slider from '../../slider/slider.js';
 import Component from '../../component.js';
 import LoadProgressBar from './load-progress-bar.js';
 import PlayProgressBar from './play-progress-bar.js';
-import SeekHandle from './seek-handle.js';
 import * as Fn from '../../utils/fn.js';
 import formatTime from '../../utils/format-time.js';
 import roundFloat from '../../utils/round-float.js';
@@ -82,11 +81,9 @@ class SeekBar extends Slider {
 SeekBar.prototype.options_ = {
   children: {
     'loadProgressBar': {},
-    'playProgressBar': {},
-    'seekHandle': {}
+    'playProgressBar': {}
   },
-  'barName': 'playProgressBar',
-  'handleName': 'seekHandle'
+  'barName': 'playProgressBar'
 };
 
 SeekBar.prototype.playerEvent = 'timeupdate';